refactor: :truck: uploadImage 위치 이동
